+++ packages/vendure-order-client/test/vendure-order-client.spec.ts
@@ -430,6 +430,14 @@ describe(
         await client.getOrderByCode(activeOrderCode);
         expect(activeOrderCode).toEqual(activeOrderStore.value.data.code);
       });
+
+      it('Poll order by code for', async () => {
+        if (!activeOrderCode) {
+          throw Error('Active order code is not defined');
+        }
+        await client.pollOrderByCode(activeOrderCode);
+        expect(activeOrderCode).toEqual(activeOrderStore.value.data.code);
+      });
     });
 
     const createNewCustomerInput: RegisterCustomerInput = {