+++ packages/vendure-plugin-invoices/src/invoice.plugin.ts
@@ -1,5 +1,4 @@
 import {
-  Logger,
   PluginCommonModule,
   RuntimeVendureConfig,
   Type,
@@ -7,31 +6,35 @@ import {
 } from '@vendure/core';
 import { AdminUiExtension } from '@vendure/ui-devkit/compiler';
 import path from 'path';
-import { InvoiceConfigEntity } from './api/entities/invoice-config.entity';
-import { InvoiceEntity } from './api/entities/invoice.entity';
+import {
+  adminSchemaExtensions,
+  shopSchemaExtensions,
+} from './api/api-extensions';
+import { InvoiceAdminResolver } from './api/invoice-admin.resolver';
+import {
+  InvoiceCommonResolver,
+  invoicePermission,
+} from './api/invoice-common.resolver';
 import { InvoiceController } from './api/invoice.controller';
-import { InvoiceResolver } from './api/invoice.resolver';
-import { InvoiceService } from './api/invoice.service';
-import { schema } from './api/schema.graphql';
-import { DataStrategy } from './api/strategies/data-strategy';
-import { DefaultDataStrategy } from './api/strategies/default-data-strategy';
-import { LocalFileStrategy } from './api/strategies/local-file-strategy';
-import { loggerCtx, PLUGIN_INIT_OPTIONS, PLUGIN_NAME } from './constants';
-import { StorageStrategy } from './index';
-import { invoicePermission } from './api/invoice.resolver';
+import { defaultLoadDataFn, LoadDataFn } from './strategies/load-data-fn';
+import { LocalFileStrategy } from './strategies/local-file-strategy';
+import { PLUGIN_INIT_OPTIONS } from './constants';
+import { InvoiceConfigEntity } from './entities/invoice-config.entity';
+import { InvoiceEntity } from './entities/invoice.entity';
+import { StorageStrategy } from './strategies/storage-strategy';
+import { InvoiceService } from './services/invoice.service';
 
 export interface InvoicePluginConfig {
   /**
-   * @deprecated We are moving this paid plugin to the Vendure Marketplace soon, so a licensekey won't be needed anymore.
-   * Existing customers will be migrated to the new system ofcourse.
-   */
-  licenseKey?: string;
-  /**
    * Hostname to use for download links, can be the Vendure instance,
-   * but also the worker instance if you want
+   * but also the worker instance if you want.
+   * Make sure to include protocol and no trailing slash, e.g. https://vendure.myshop.com
    */
   vendureHost: string;
-  dataStrategy: DataStrategy;
+  /**
+   * Load custom data that is passed in to your HTML/handlebars template
+   */
+  loadDataFn: LoadDataFn;
   storageStrategy: StorageStrategy;
 }
 
@@ -44,8 +47,12 @@ export interface InvoicePluginConfig {
   ],
   controllers: [InvoiceController],
   adminApiExtensions: {
-    schema: schema as any,
-    resolvers: [InvoiceResolver],
+    schema: adminSchemaExtensions,
+    resolvers: [InvoiceAdminResolver, InvoiceCommonResolver],
+  },
+  shopApiExtensions: {
+    schema: shopSchemaExtensions,
+    resolvers: [InvoiceCommonResolver],
   },
   compatibility: '^2.0.0',
   configuration: (config: RuntimeVendureConfig) => {
@@ -60,9 +67,9 @@ export class InvoicePlugin {
     config: Partial<InvoicePluginConfig> & { vendureHost: string }
   ): Type<InvoicePlugin> {
     InvoicePlugin.config = {
-      ...config,
+      vendureHost: config.vendureHost,
       storageStrategy: config.storageStrategy || new LocalFileStrategy(),
-      dataStrategy: config.dataStrategy || new DefaultDataStrategy(),
+      loadDataFn: config.loadDataFn || defaultLoadDataFn,
     };
     return this;
   }