+++ packages/vendure-plugin-invoices/src/strategies/local-file-strategy.ts
@@ -2,7 +2,7 @@ import { Response } from 'express';
 import { createReadStream, promises as fs, ReadStream } from 'fs';
 import path from 'path';
 import { InvoiceEntity } from '../entities/invoice.entity';
-import { exists, zipFiles, ZippableFile } from '../file.util';
+import { exists } from '../util/file.util';
 import { LocalStorageStrategy } from './storage-strategy';
 
 /**
@@ -14,30 +14,22 @@ export class LocalFileStrategy implements LocalStorageStrategy {
 
   async init(): Promise<void> {}
 
-  async save(tmpFile: string, invoiceNumber: number, channelToken: string) {
+  async save(
+    tmpFile: string,
+    invoiceNumber: number,
+    channelToken: string,
+    isCreditInvoice: boolean
+  ) {
     if (!(await exists(this.invoiceDir))) {
       await fs.mkdir(this.invoiceDir);
     }
-    const fileName = path.basename(tmpFile);
-    const newPath = `${this.invoiceDir}/${invoiceNumber}.pdf`;
+    let name = `${invoiceNumber}.pdf`;
+    if (isCreditInvoice) {
+      name = `${invoiceNumber}-credit.pdf`;
+    }
+    const newPath = `${this.invoiceDir}/${name}`;
     await fs.rename(tmpFile, newPath);
     return newPath;
-  }
-
-  async streamMultiple(
-    invoices: InvoiceEntity[],
-    res: Response
-  ): Promise<ReadStream> {
-    res.set({
-      'Content-Type': 'application/zip',
-      'Content-Disposition': `inline; filename="invoices-${invoices.length}.zip"`,
-    });
-    const zippableFiles: ZippableFile[] = invoices.map((invoice) => ({
-      path: invoice.storageReference,
-      name: invoice.invoiceNumber + '.pdf',
-    }));
-    const zipFile = await zipFiles(zippableFiles);
-    return createReadStream(zipFile);
   }
 
   async streamFile(invoice: InvoiceEntity, res: Response): Promise<ReadStream> {