+++ packages/vendure-plugin-invoices/src/ui/invoices.module.ts
@@ -2,6 +2,7 @@ import { NgModule } from '@angular/core';
 import { RouterModule } from '@angular/router';
 import { SharedModule } from '@vendure/admin-ui/core';
 import { InvoicesComponent } from './invoices.component';
+import { InvoiceDetailViewComponent } from './invoices-detail-view/invoices-detail-view.component';
 
 @NgModule({
   imports: [
@@ -16,6 +17,6 @@ import { InvoicesComponent } from './invoices.component';
     ]),
   ],
   providers: [],
-  declarations: [InvoicesComponent],
+  declarations: [InvoicesComponent, InvoiceDetailViewComponent],
 })
 export class InvoicesModule {}