+++ packages/vendure-plugin-invoices/src/util/file.util.ts
@@ -1,6 +1,7 @@
+import { Logger } from '@vendure/core';
+import fs from 'fs/promises';
 import * as tmp from 'tmp';
-import { promises as fs, createWriteStream, createReadStream } from 'fs';
-import AdmZip = require('adm-zip');
+import { loggerCtx } from '../constants';
 
 export async function createTempFile(postfix: string): Promise<string> {
   return new Promise((resolve, reject) => {
@@ -14,21 +15,6 @@ export async function createTempFile(postfix: string): Promise<string> {
   });
 }
 
-export interface ZippableFile {
-  name: string;
-  path: string;
-}
-
-export async function zipFiles(files: ZippableFile[]): Promise<string> {
-  const zip = new AdmZip();
-  for (const file of files) {
-    zip.addLocalFile(file.path, undefined, file.name);
-  }
-  const tmpFilePath = await createTempFile('.zip');
-  await zip.writeZip(tmpFilePath);
-  return tmpFilePath;
-}
-
 export async function exists(path: string): Promise<boolean> {
   let exists = false;
   try {
@@ -38,4 +24,17 @@ export async function exists(path: string): Promise<boolean> {
     exists = false;
   }
   return exists;
+}
+
+/**
+ * Attempt deletion of file, but swallow any errors.
+ */
+export function safeRemove(path: string): void {
+  fs.unlink(path).catch((err: any | undefined) => {
+    Logger.error(
+      `Could not remove file ${path}: ${err?.message}`,
+      loggerCtx,
+      err?.stack
+    );
+  });
 }